+++ apps/wallet/src/components/amounts-display/ErgAmount.tsx
@@ -3,6 +3,7 @@ import TokenAmountDisplay from './TokenAmountDisplay';
 interface ErgAmountDisplayPropsType {
   amount: bigint;
   displayDecimal?: number;
+  forceDisplay?: boolean;
 }
 
 const ErgAmountDisplay = (props: ErgAmountDisplayPropsType) => {
@@ -11,6 +12,7 @@ const ErgAmountDisplay = (props: ErgAmountDisplayPropsType) => {
       amount={props.amount}
       decimal={9}
       displayDecimal={props.displayDecimal}
+      forceDisplay={props.forceDisplay}
     />
   );
 };