+++ packages/electron/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @minotaur-ergo/electron
 
+## 0.2.0
+
+### Minor Changes
+
+- Add cli directory.
+
 ## 0.1.0
 
 ### Minor changes